move to vector_dictionary
