+++ cpp/src/arrow/compute/kernels/CMakeLists.txt
@@ -95,6 +95,7 @@ add_arrow_compute_test(vector_test
                        vector_replace_test.cc
                        vector_run_end_encode_test.cc
                        select_k_test.cc
+                       vector_dictionary_test.cc
                        EXTRA_LINK_LIBS
                        arrow_compute_kernels_testing)
