+++ cpp/src/arrow/compute/kernels/vector_hash.cc
@@ -795,167 +795,6 @@ class DictionaryDecodeMetaFunction : public MetaFunction {
     }
   }
 };
-
-// Dictionary compaction implementation
-
-const FunctionDoc dictionary_compaction_doc{
-    "Compact dictionary array",
-    ("Return a compacted version of the dictionary array input,\n"
-     "which would remove unused values in dictionary."),
-    {"dictionary_array"}};
-
-class DictionaryCompactionKernel : public KernelState {
- public:
-  virtual Result<std::shared_ptr<Array>> Exec(std::shared_ptr<Array> dict_array,
-                                              ExecContext* ctx) const = 0;
-};
-
-template <typename IndiceArrowType>
-class DictionaryCompactionKernelImpl : public DictionaryCompactionKernel {
-  using BuilderType = NumericBuilder<IndiceArrowType>;
-  using CType = typename IndiceArrowType::c_type;
-
- public:
-  Result<std::shared_ptr<Array>> Exec(std::shared_ptr<Array> dict_array,
-                                      ExecContext* ctx) const override {
-    const DictionaryArray& casted_dict_array =
-        checked_cast<const DictionaryArray&>(*dict_array);
-
-    const std::shared_ptr<Array>& dict = casted_dict_array.dictionary();
-    if (dict->length() == 0) {
-      return dict_array;
-    }
-
-    const std::shared_ptr<Array>& indice = casted_dict_array.indices();
-    if (indice->length() == 0) {
-      ARROW_ASSIGN_OR_RAISE(auto empty_dict,
-                            MakeEmptyArray(dict->type(), ctx->memory_pool()));
-      return DictionaryArray::FromArrays(dict_array->type(), indice, empty_dict);
-    }
-    const CType* indices_data =
-        reinterpret_cast<const CType*>(indice->data()->buffers[1]->data());
-
-    // check whether the input is compacted
-    std::vector<bool> dict_used(dict->length(), false);
-    int64_t dict_used_count = 0;
-    int64_t offset = indice->data()->offset;
-    for (int64_t i = 0; i < indice->length(); i++) {
-      if (indice->IsNull(i)) {
-        continue;
-      }
-
-      CType cur_indice = indices_data[i + offset];
-      if (!dict_used[cur_indice]) {
-        dict_used[cur_indice] = true;
-        dict_used_count++;
-
-        if (dict_used_count == dict->length()) {  // input is already compacted
-          return dict_array;
-        }
-      }
-    }
-
-    // dictionary compaction
-    std::shared_ptr<arrow::Array> compacted_dict;
-    if (dict_used_count == 0) {
-      ARROW_ASSIGN_OR_RAISE(compacted_dict,
-                            MakeEmptyArray(dict->type(), ctx->memory_pool()));
-    } else {
-      std::vector<CType> dict_indice;
-      CType len = (CType)dict->length();
-      for (CType i = 0; i < len; i++) {
-        if (dict_used[i]) {
-          dict_indice.push_back(i);
-        }
-      }
-      BuilderType dict_indice_builder;
-      ARROW_RETURN_NOT_OK(dict_indice_builder.AppendValues(dict_indice));
-      ARROW_ASSIGN_OR_RAISE(std::shared_ptr<arrow::Array> compacted_dict_indices,
-                            dict_indice_builder.Finish());
-      ARROW_ASSIGN_OR_RAISE(
-          auto compacted_dict_res,
-          Take(dict, compacted_dict_indices, TakeOptions::NoBoundsCheck(), ctx));
-      compacted_dict = compacted_dict_res.make_array();
-    }
-
-    // indice changes
-    std::vector<CType> indice_minus_number(dict->length(), 0);
-    if (!dict_used[0]) {
-      indice_minus_number[0] = 1;
-    }
-    for (int64_t i = 1; i < dict->length(); i++) {
-      indice_minus_number[i] = indice_minus_number[i - 1];
-      if (!dict_used[i]) {
-        indice_minus_number[i] = indice_minus_number[i] + 1;
-      }
-    }
-
-    std::vector<CType> raw_changed_indice(indice->length(), 0);
-    std::vector<bool> is_valid(indice->length(), true);
-    for (int64_t i = 0; i < indice->length(); i++) {
-      if (indice->IsNull(i)) {
-        is_valid[i] = false;
-      } else {
-        CType cur_indice = indices_data[i + offset];
-        raw_changed_indice[i] = cur_indice - indice_minus_number[cur_indice];
-      }
-    }
-    BuilderType indice_builder;
-    if (indice->null_count() != 0) {
-      ARROW_RETURN_NOT_OK(indice_builder.AppendValues(raw_changed_indice, is_valid));
-    } else {
-      ARROW_RETURN_NOT_OK(indice_builder.AppendValues(raw_changed_indice));
-    }
-    ARROW_ASSIGN_OR_RAISE(std::shared_ptr<arrow::Array> changed_indice,
-                          indice_builder.Finish());
-
-    ARROW_ASSIGN_OR_RAISE(
-        auto res,
-        DictionaryArray::FromArrays(dict_array->type(), changed_indice, compacted_dict));
-    return res;
-  }
-};
-
-Result<std::unique_ptr<KernelState>> DictionaryCompactionInit(
-    KernelContext* ctx, const KernelInitArgs& args) {
-  const auto& dict_type =
-      checked_cast<const DictionaryType&>(*(args.inputs[0].owned_type));
-  switch (dict_type.index_type()->id()) {
-    case Type::UINT8:
-      return std::make_unique<DictionaryCompactionKernelImpl<UInt8Type>>();
-    case Type::INT8:
-      return std::make_unique<DictionaryCompactionKernelImpl<Int8Type>>();
-    case Type::UINT16:
-      return std::make_unique<DictionaryCompactionKernelImpl<UInt16Type>>();
-    case Type::INT16:
-      return std::make_unique<DictionaryCompactionKernelImpl<Int16Type>>();
-    case Type::UINT32:
-      return std::make_unique<DictionaryCompactionKernelImpl<UInt32Type>>();
-    case Type::INT32:
-      return std::make_unique<DictionaryCompactionKernelImpl<Int32Type>>();
-    case Type::UINT64:
-      return std::make_unique<DictionaryCompactionKernelImpl<UInt64Type>>();
-    case Type::INT64:
-      return std::make_unique<DictionaryCompactionKernelImpl<Int64Type>>();
-    default:
-      ARROW_CHECK(false) << "unreachable";
-      return Status::TypeError("Expected an Indice Type of Int or UInt");
-  }
-}
-
-Status DictionaryCompactionExec(KernelContext* ctx, const ExecSpan& batch,
-                                ExecResult* out) {
-  if (batch[0].is_scalar()) {
-    return Status::TypeError("Expected an Array or a Chunked Array");
-  }
-
-  const DictionaryCompactionKernel& Kernel =
-      checked_cast<const DictionaryCompactionKernel&>(*ctx->state());
-  ARROW_ASSIGN_OR_RAISE(std::shared_ptr<Array> compacted_dict_array,
-                        Kernel.Exec(batch[0].array.ToArray(), ctx->exec_context()));
-  out->value = compacted_dict_array->data();
-  return Status::OK();
-}
 }  // namespace
 
 void RegisterVectorHash(FunctionRegistry* registry) {
@@ -1015,18 +854,6 @@ void RegisterVectorHash(FunctionRegistry* registry) {
 
 void RegisterDictionaryDecode(FunctionRegistry* registry) {
   DCHECK_OK(registry->AddFunction(std::make_shared<DictionaryDecodeMetaFunction>()));
-}
-
-void RegisterDictionaryCompaction(FunctionRegistry* registry) {
-  VectorKernel base;
-  base.init = DictionaryCompactionInit;
-  base.exec = DictionaryCompactionExec;
-  base.signature = KernelSignature::Make({Type::DICTIONARY}, FirstType);
-
-  auto dictionary_compaction = std::make_shared<VectorFunction>(
-      "dictionary_compaction", Arity::Unary(), dictionary_compaction_doc);
-  DCHECK_OK(dictionary_compaction->AddKernel(base));
-  DCHECK_OK(registry->AddFunction(std::move(dictionary_compaction)));
 }
 
 }  // namespace internal
